feat: add detection of Raspberry Pi model to ContinuousServo class
The whole 'servo jitter' problem is its own can of worms. The Pi4 can handle it through PiGPIOFactory, because developers of its native pigpio library figured out a way to use the chip's architecture for hardware PWM. The Pi 5 doesn't use pigpio due to the new chip architecture, so hardware PWM is not possible at the mo, and RPiGPIOFactory must be used.